clean up CLI
