+++ crates/runtime-substrate/src/lib.rs
@@ -35,6 +35,7 @@
 
 pub mod boundary;
 pub mod chain;
+pub mod content_address;
 pub mod cross_check;
 pub mod error;
 pub mod facade;
@@ -56,6 +57,7 @@ pub mod types;
 
 pub use boundary::{check_call_method, check_run_script};
 pub use chain::ChainAccessor;
+pub use content_address::{RuntimeScriptIdentity, RUNTIME_SCRIPT_IRI_PREFIX};
 pub use cross_check::{
     is_cross_check_failure, prepare_substrate_side, verify_in_worker, CrossCheckError,
     CrossCheckOutcome, ProvenanceDirAction, SubstratePrepareError, DEFAULT_PROVENANCE_DIR,
